Significantly improved the speed of the program
By using more efficient libraries and testing what would be most optimal in this use case, I was able to utilize NumPy's method for shuffling array's to improve the speed at which arrays are shuffled.

As well as that, I was able to come up with a more efficient way to iterate through the food sources so that there are less random numbers that need to be generated, essentially just generating a list with every index of the food sources twice, since each food source can support 2 creatures, and then shuffling it to create the new array, and simply iterating through that and the creature array to assign them to their positions improved that function's efficiency.

The reproduce function was the next bottleneck I had to look at and by simply reducing its complexity and the amount of things it had to do, I was able to slightly improve its efficiency.

So even though the last version struggled at numbers beyond 100,000 for both the creatures and food source as the starting condition, with generations taking around a second to go through, I now don't reach that point until around 400,000 now.

I am trying to avoid making this a parallel program since that seems to defeat the point of this project, and is not normally natively supported in python, but I may make a version of it that is parallel using some python libraries and see how far I can push that one.