Minor changes that don't have huge performance impact.
Changes to the iteration of adding creatures to the food sources that should be equivalent but don't require checking if the creature count has exceeded the capacity of all food sources. Also moved the resetting of food sources into the reproduction loop to prevent iterating over the food sources multiple times. This might have slightly sped up the program, but nothing too significant I noticed.